Remove cards directory from home and move up a level.
