+++ myuw/static/vue/components/home/events/list-events.vue
