+++ myuw/static/vue/components/home/new_student/thank-you.vue
