+++ myuw/static/vue/components/home/notices.vue
