+++ myuw/static/vue/components/home/registration/resources.vue
