+++ myuw/static/vue/components/home/registration/status.vue
