+++ myuw/static/vue/tests/thank-you.test.js
@@ -3,7 +3,7 @@ import {mount, shallowMount, createLocalVue} from '@vue/test-utils';
 import BootstrapVue from 'bootstrap-vue';
 import Vuex from 'vuex';
 import notices from '../vuex/store/notices';
-import ThankYouCard from '../components/home/cards/new_student/thank-you.vue';
+import ThankYouCard from '../components/home/new_student/thank-you.vue';
 
 import mockNotices from './mock_data/notices.json';
